Refactor cog creation script and use band tags Add option to use metadata table to set as band tags and refactors file
If the .csv is provided it will be used to filter which bands are used
and the order. This commit also gets rid of the rio_cogeo cog contructor
because with the "new" COG driver of GDAL it is not needed anymore. ALso
I tried to reorganize a bit the file with relative luck.