do not attempt to reset tasks to placeholder container images
This is a simple way to avoid the chicken-and-egg problem with
tasks needing to reference the real foodscapes OCI images while there's
not even an ECR registry set up for these yet.

In practice we only need to set these tasks up when we create the ECS
cluster, and future updates are all handled by the deployment setup in
GitHub actions.